requirements.txt
Django==1.10.5
django-filter==1.0.1
djangorestframework==3.5.3
gunicorn==19.6.0
Markdown==2.6.7
mysqlclient==1.3.9
olefile==0.44
Pillow==4.0.0
